3_real_life_application/lsc-1-people-detection/requirements.txt
parliament-functions==0.1.0
boto3
cython 
pillow>=7.0.0 
numpy>=1.18.1 
opencv-python>=4.1.2 
pytest==7.1.3 
tqdm==4.64.1 
scipy>=1.7.3 
matplotlib>=3.4.3 
mock==4.0.3
imageai
--extra-index-url https://download.pytorch.org/whl/cpu 
torch>=1.9.0 
torchvision>=0.10.0